+++ packages/mrm-task-package/Changelog.md
@@ -1,3 +1,8 @@
+## Next
+
+* New options: minNode, license.
+* Update engines.node on second run.
+
 ## 1.0.1
 
 * Fix: Correct package name.